Builder: Update AtomX script to 3.0

+++ AtomX.sh
@@ -4,60 +4,29 @@
 KBUILD_USER="AtomX"
 KBUILD_HOST="Drone"
 
-# Verbose mode (verbose build: 0 | silent build: 1) 
-# (default: 0 | modes: 0, 1)
-SILENCE='0'
-
-# Do modules (Perform modules post compile process: 1 | No modules: 0)
-# (default: 0 | modes: 0, 1)
-MODULES='1'
-
 # Build type (Fresh build: clean | incremental build: dirty)
 # (default: dirty | modes: clean, dirty)
 BUILD='clean'
-
-# Compiler
-# (default: clang | modes: clang, gcc) 
-COMPILER='clang'
-
-#  DEVICES CONFIG  #
-
-# Device Name
-# Eg: (For lisa: Xiaomi 11 lite 5G NE) 
-DEVICENAME='Xiaomi 11 lite 5G NE'
-
-# Device Codename
-# Eg: (For lisa: Xiaomi 11 lite 5G NE) 
-CODENAME='lisa'
-
-# Defconfig Suffix
-# Eg: (For perf defconfig: perf, For QGKI defconfig: qgki) 
-SUFFIX='qgki'
-
-# Target Image
-# Eg: (image: Image, Image.gz, Image.gz-dtb) 
-TARGET='Image'
-
 ############################################################################
 
 ########################    DIRECTOR PATHS   ###############################
 
-# Kernel directory
+# Kernel Directory
 KERNEL_DIR=`pwd`
 
-# PATH (default paths may not work!)
+# Propriatary Directory (default paths may not work!)
 PRO_PATH="$KERNEL_DIR/.."
 
-# Anykernel directories
+# Anykernel Directories
 AK3_DIR="$PRO_PATH/AnyKernel3"
 AKSH="$AK3_DIR/anykernel.sh"
 AKVDR="$AK3_DIR/modules/vendor/lib/modules"
 
-# Toolchain directory
+# Toolchain Directory
 TLDR="$PRO_PATH/toolchains"
 
-# DTS path
-DTS_PATH="$KERNEL_DIR/work/arch/arm64/boot/dts/vendor/qcom"
+# Device Tree Blob Directory
+DTB_PATH="$KERNEL_DIR/work/arch/arm64/boot/dts/vendor/qcom"
 
 ############################################################################
 
@@ -86,6 +55,14 @@ muke() {
 	make $@ $KERN_MAKE_ARGS
 }
 
+usage() {
+	inform " ./AtomX.sh <arg>
+		--compiler   sets the compiler to be used
+		--device     sets the device for kernel build
+		--silence    Silence shell output of Kbuild"
+	exit 2
+}
+
 ############################################################################
 
 compiler_setup() {
@@ -93,30 +70,26 @@ compiler_setup() {
 	case $COMPILER in
 		clang)
 			CC='clang'
-			HOSTCC="$CC"
-			HOSTCXX="$CC++"
-			CC_64='aarch64-linux-gnu-'
 			C_PATH="$TLDR/clang"
 		;;
 		gcc)
-			HOSTCC='gcc'
-			CC_64='aarch64-elf-'
 			CC='aarch64-elf-gcc'
-			HOSTCXX='aarch64-elf-g++'
+			KERN_MAKE_ARGS="                 \
+					HOSTCC=gcc                   \
+					CC=$CC                       \
+					HOSTCXX=aarch64-elf-g++      \
+					CROSS_COMPILE=aarch64-elf-"
 			C_PATH="$TLDR/gcc-arm64"
 		;;
 	esac
 	CC_32="$TLDR/gcc-arm/bin/arm-eabi-"
 	CC_COMPAT="$TLDR/gcc-arm/bin/arm-eabi-gcc"
 
-	KERN_MAKE_ARGS="O=work ARCH=arm64    \
-		CC=$CC                           \
+	KERN_MAKE_ARGS="$KERN_MAKE_ARGS    \
 		LLVM=1                           \
 		LLVM_IAS=1                       \
 		HOSTLD=ld.lld                    \
-		HOSTCC=$HOSTCC                   \
-		HOSTCXX=$HOSTCXX                 \
-		CROSS_COMPILE=$CC_64             \
+		O=work ARCH=arm64                \
 		CC_COMPAT=$CC_COMPAT             \
 		DTC_EXT=$(which dtc)             \
 		PATH=$C_PATH/bin:$PATH           \
@@ -127,8 +100,13 @@ compiler_setup() {
 ############################################################################
 }
 
-build() {
+kernel_builder() {
 ##################################  BUILD  #################################
+	if [[ -z $CODENAME ]]; then
+		error 'device not mentioned'
+		exit 1
+	fi
+
 	case $BUILD in
 		clean)
 			rm -rf work || mkdir work
@@ -140,20 +118,20 @@ build() {
 		;;
 	esac
 
+	# Build Start
+	BUILD_START=$(date +"%s")
+
 	DFCF="vendor/${CODENAME}-${SUFFIX}_defconfig"
 
 	inform "
-  *************Build Triggered*************
-  Date: <code>$(date +"%Y-%m-%d %H:%M")</code>
-  Build Type: <code>$BUILD_TYPE</code>
-  Device: <code>$DEVICENAME</code>
-  Codename: <code>$CODENAME</code>
-  Compiler: <code>$($C_PATH/bin/$CC --version | head -n 1)</code>
-  Compiler_32: <code>$($CC_COMPAT --version | head -n 1)</code>
-  "
-
-	# Build Start
-	BUILD_START=$(date +"%s")
+		*************Build Triggered*************
+		Date: <code>$(date +"%Y-%m-%d %H:%M")</code>
+		Build Type: <code>$BUILD_TYPE</code>
+		Device: <code>$DEVICENAME</code>
+		Codename: <code>$CODENAME</code>
+		Compiler: <code>$($C_PATH/bin/$CC --version | head -n 1 | perl -pe 's/\(http.*?\)//gs')</code>
+		Compiler_32: <code>$($CC_COMPAT --version | head -n 1)</code>
+	"
 
 	# Make .config
 	muke $DFCF
@@ -162,9 +140,10 @@ build() {
 	muke -j$(nproc)
 
 	if [[ "$MODULES" == "1" ]]; then
-		muke -j$(nproc) 'modules_install' \
-						INSTALL_MOD_STRIP=1 \
-						INSTALL_MOD_PATH="modules"
+		muke -j$(nproc)        \
+			'modules_install'    \
+			INSTALL_MOD_STRIP=1  \
+			INSTALL_MOD_PATH="modules"
 	fi
 
 	# Build End
@@ -173,20 +152,32 @@ build() {
 	DIFF=$(($BUILD_END - $BUILD_START))
 
 	if [[ -f $KERNEL_DIR/work/arch/arm64/boot/$TARGET ]]; then
-		zip_ak
+		zipper
 	else
 		error 'Kernel image not found'
 	fi
 ############################################################################
 }
 
-zip_ak() {
+zipper() {
 ####################################  ZIP  #################################
+	source work/.config
+
+	VERSION=`echo $CONFIG_LOCALVERSION | cut -c 8-`
+	KERNEL_VERSION=$(make kernelversion)
+	LAST_COMMIT=$(git show -s --format=%s)
+	LAST_HASH=$(git rev-parse --short HEAD)
+
 	if [[ ! -d $AK3_DIR ]]; then
 		error 'Anykernel not present cannot zip'
 	fi
+	if [[ ! -d "$KERNEL_DIR/out" ]]; then
+		mkdir $KERNEL_DIR/out
+	fi
 
 	cp $KERNEL_DIR/work/arch/arm64/boot/$TARGET $AK3_DIR
+	cp $DTB_PATH/*.dtb $AK3_DIR/dtb
+	cp $DTB_PATH/*.img $AK3_DIR/
 	if [[ "$MODULES" == "1" ]]; then
 		MOD_NAME="$(cat work/include/generated/utsrelease.h | cut -c 21- | tr -d '"')"
 		MOD_PATH="work/modules/lib/modules/$MOD_NAME"
@@ -196,36 +187,30 @@ zip_ak() {
 		cp $MOD_PATH/modules.order $AKVDR/modules.load
 		sed -i 's/\(kernel\/[^: ]*\/\)\([^: ]*\.ko\)/\/vendor\/lib\/modules\/\2/g' $AKVDR/modules.dep
 		sed -i 's/.*\///g' $AKVDR/modules.load
-		cp $DTS_PATH/*.dtb $AK3_DIR/dtb
-		cp $DTS_PATH/*.img $AK3_DIR/
 	fi
-
-	KERNEL_VERSION=$(make kernelversion)
-	LAST_COMMIT=$(git show -s --format=%s)
-	LAST_HASH=$(git rev-parse --short HEAD)
 
 	cd $AK3_DIR
 
-	make zip
+	make zip VERSION=$VERSION
 
+	inform "
+		*************AtomX-Kernel*************
+		Linux Version: <code>$KERNEL_VERSION</code>
+		AtomX-Version: <code>$VERSION</code>
+		CI: <code>$KBUILD_HOST</code>
+		Core count: <code>$(nproc)</code>
+		Compiler: <code>$($C_PATH/bin/$CC --version | head -n 1 | perl -pe 's/\(http.*?\)//gs')</code>
+		Compiler_32: <code>$($CC_COMPAT --version | head -n 1)</code>
+		Device: <code>$DEVICENAME</code>
+		Codename: <code>$CODENAME</code>
+		Build Date: <code>$(date +"%Y-%m-%d %H:%M")</code>
+		Build Type: <code>$BUILD_TYPE</code>
 
-inform "
-  *************AtomX-Kernel*************
-  Linux Version: <code>$KERNEL_VERSION</code>
-  CI: <code>$KBUILD_HOST</code>
-  Core count: <code>$(nproc)</code>
-  Compiler: <code>$($C_PATH/bin/$CC --version | head -n 1)</code>
-  Compiler_32: <code>$($CC_COMPAT --version | head -n 1)</code>
-  Device: <code>$DEVICENAME</code>
-  Codename: <code>$CODENAME</code>
-  Build Date: <code>$(date +"%Y-%m-%d %H:%M")</code>
-  Build Type: <code>$BUILD_TYPE</code>
+		-----------last commit details-----------
+		Last commit (name): <code>$LAST_COMMIT</code>
 
-  -----------last commit details-----------
-  Last commit (name): <code>$LAST_COMMIT</code>
-
-  Last commit (hash): <code>$LAST_HASH</code>
-  "
+		Last commit (hash): <code>$LAST_HASH</code>
+	"
 
 	telegram-send --file *-signed.zip
 
@@ -238,9 +223,55 @@ inform "
 ############################################################################
 }
 
+###############################  COMMAND_MODE  ##############################
+if [[ -z $* ]]; then
+	usage
+fi
+
+for arg in "$@"; do
+	case "${arg}" in
+		"--compiler="*)
+			COMPILER=${arg#*=}
+			case ${COMPILER} in
+				clang)
+					COMPILER="clang"
+				;;
+				gcc)
+					COMPILER="gcc"
+				;;
+				*)
+					usage
+				;;
+			esac
+		;;
+		"--device="*)
+			CODENAME=${arg#*=}
+			case $CODENAME in
+				lisa)
+					DEVICENAME='Xiaomi 11 lite 5G NE'
+					CODENAME='lisa'
+					SUFFIX='qgki'
+					MODULES='1'
+					TARGET='Image'
+				;;
+				*)
+					error 'device not supported'
+				;;
+			esac
+		;;
+		"--silence")
+			SILENCE='1'
+		;;
+		*)
+			usage
+		;;
+	esac
+done
+############################################################################
+
 # Remove testing of System.map as test always fails to check for file
 # DO NOT MODIFY!!!!
 sed -i '13d;14d;15d;16d;17d' $KERNEL_DIR/scripts/depmod.sh
 
 compiler_setup
-build
+kernel_builder
